Study the AOP.
Found out the reason that why my annotation base pointcut not working.
As said in the Spring doc:
The key thing to understand here is that the client code inside the main(..) method of the Main class has a reference to the proxy. This means that method calls on that object reference are calls on the proxy. As a result, the proxy can delegate to all of the interceptors (advice) that are relevant to that particular method call. However, once the call has finally reached the target object (the SimplePojo, reference in this case), any method calls that it may make on itself, such as this.bar() or this.foo(), are going to be invoked against the this reference, and not the proxy. This has important implications. It means that self-invocation is not going to result in the advice associated with a method invocation getting a chance to execute.

So the following example:
class YourService {
    public void outter {
        inner();
    }

    @YourAnnocation
    public void inner() {
    }
}

If you call outter(), your AOP advice won't be called.