Design the layout in Home components.
Flexbox seems to be the best overall layout strategy. However we need to deal with the last rows alignment with dynamic number of Apps. We append
additional elements to the last row, it seems to be the best solution.
But we still need to figuer out how many additional elements to be
added. Right now we just add as many as we think it would be enough.